Update acpick
Some improvements.  The scope for the arrays for track names and card names is fixed.

The script allows choice of track name (and car) and will rebuild the server_cfg.ini file using that chosen track.

Note that it only goes through the motion of choosing a car.  It turns out the car selection is more complicated and will require more thought.  The car names put into server_cfg.ini must match entries in the entry_list.ini file which describes individual cars.  So, this will need modification to build both files.

Also TO-DO -- On further testing it appears track names with a blank CONFIG_TRACK value are only valid if there is no other entry for that same track with a non-blank CONFIG_TRACK setting.  Therefore, the script needs to be modified to eliminate the blank CONFIG_TRACK entries when a subsequent non-blank CONFIG track is added.